[FEAT]일반챌린지참가자저장
Former-commit-id: f462dee47cdc6a3f5b4dc32656cf05646bba69d5

+++ src/main/java/com/ssafy/ChallenMungs/challenge/general/entity/GeneralParticipant.java
@@ -18,8 +18,8 @@ public class GeneralParticipant {
     String loginId;
 
     @Column(name = "challenge_id")
-    int challengeId;
+    Long challengeId;
 
     @Column(name = "success_count")
-    int successCount;
+    Integer successCount;
 }
